requirements.txt
fastapi
uvicorn
transformers
torch
nltk
aiofiles
python-multipart
PyMuPDF
python-docx
protobuf